requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.2.3
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
inflection==0.5.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.6.0
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
praw==7.1.0
prawcore==1.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
Quandl==3.6.0
requests==2.25.1
retrying==1.3.3
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.75.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing==3.7.4.3
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.3
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
yfinance==0.1.55
